refactor: improve attendance page structure and components, add table skeleton
